+++ dependencies/predis/predis/src/ClientInterface.php
@@ -3,7 +3,8 @@
 /*
  * This file is part of the Predis package.
  *
- * (c) Daniele Alessandri <suppakilla@gmail.com>
+ * (c) 2009-2020 Daniele Alessandri
+ * (c) 2021-2023 Till Krüss
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,9 +14,11 @@ namespace Predis;
 
 use Predis\Command\Argument\Geospatial\ByInterface;
 use Predis\Command\Argument\Geospatial\FromInterface;
+use Predis\Command\Argument\Server\LimitOffsetCount;
 use Predis\Command\Argument\Server\To;
 use Predis\Command\CommandInterface;
 use Predis\Command\FactoryInterface;
+use Predis\Command\Redis\Container\FunctionContainer;
 use Predis\Configuration\OptionsInterface;
 use Predis\Connection\ConnectionInterface;
 use Predis\Response\Status;
@@ -32,8 +35,9 @@ use Predis\Response\Status;
  * @method int               del(string[]|string $keyOrKeys, string ...$keys = null)
  * @method string|null       dump(string $key)
  * @method int               exists(string $key)
- * @method int               expire(string $key, int $seconds)
- * @method int               expireat(string $key, int $timestamp)
+ * @method int               expire(string $key, int $seconds, string $expireOption = '')
+ * @method int               expireat(string $key, int $timestamp, string $expireOption = '')
+ * @method int               expiretime(string $key)
  * @method array             keys(string $pattern)
  * @method int               move(string $key, int $db)
  * @method mixed             object($subcommand, string $key)
@@ -46,13 +50,14 @@ use Predis\Response\Status;
  * @method int               renamenx(string $key, string $target)
  * @method array             scan($cursor, array $options = null)
  * @method array             sort(string $key, array $options = null)
+ * @method array             sort_ro(string $key, ?string $byPattern = null, ?LimitOffsetCount $limit = null, array $getPatterns = [], ?string $sorting = null, bool $alpha = false)
  * @method int               ttl(string $key)
  * @method mixed             type(string $key)
  * @method int               append(string $key, $value)
- * @method int               bitcount(string $key, $start = null, $end = null)
+ * @method int               bitcount(string $key, $start = null, $end = null, string $index = 'byte')
  * @method int               bitop($operation, $destkey, $key)
  * @method array|null        bitfield(string $key, $subcommand, ...$subcommandArg)
- * @method int               bitpos(string $key, $bit, $start = null, $end = null)
+ * @method int               bitpos(string $key, $bit, $start = null, $end = null, string $index = 'byte')
  * @method array             blmpop(int $timeout, array $keys, string $modifier = 'left', int $count = 1)
  * @method array             bzpopmax(array $keys, int $timeout)
  * @method array             bzpopmin(array $keys, int $timeout)
@@ -60,6 +65,7 @@ use Predis\Response\Status;
  * @method int               decr(string $key)
  * @method int               decrby(string $key, int $decrement)
  * @method Status            failover(?To $to = null, bool $abort = false, int $timeout = -1)
+ * @method mixed             fcall(string $function, array $keys, ...$args)
  * @method string|null       get(string $key)
  * @method int               getbit(string $key, $offset)
  * @method int|null          getex(string $key, $modifier = '', $value = false)
@@ -99,6 +105,7 @@ use Predis\Response\Status;
  * @method array|null        blpop(array|string $keys, int|float $timeout)
  * @method array|null        brpop(array|string $keys, int|float $timeout)
  * @method string|null       brpoplpush(string $source, string $destination, int|float $timeout)
+ * @method mixed             lcs(string $key1, string $key2, bool $len = false, bool $idx = false, int $minMatchLen = 0, bool $withMatchLen = false)
  * @method string|null       lindex(string $key, int $index)
  * @method int               linsert(string $key, $whence, $pivot, $value)
  * @method int               llen(string $key)
@@ -120,6 +127,7 @@ use Predis\Response\Status;
  * @method string[]          sdiff(array|string $keys)
  * @method int               sdiffstore(string $destination, array|string $keys)
  * @method string[]          sinter(array|string $keys)
+ * @method int               sintercard(array $keys, int $limit = 0)
  * @method int               sinterstore(string $destination, array|string $keys)
  * @method int               sismember(string $key, string $member)
  * @method string[]          smembers(string $key)
@@ -170,6 +178,7 @@ use Predis\Response\Status;
  * @method array             zrevrangebylex(string $key, string $start, string $stop, array $options = null)
  * @method int               zremrangebylex(string $key, string $min, string $max)
  * @method int               zlexcount(string $key, string $min, string $max)
+ * @method int               pexpiretime(string $key)
  * @method int               pfadd(string $key, array $elements)
  * @method mixed             pfmerge(string $destinationKey, array|string $sourceKeys)
  * @method int               pfcount(string[]|string $keyOrKeys, string ...$keys = null)
@@ -181,7 +190,9 @@ use Predis\Response\Status;
  * @method mixed             unwatch()
  * @method mixed             watch(string $key)
  * @method mixed             eval(string $script, int $numkeys, string ...$keyOrArg = null)
+ * @method mixed             eval_ro(string $script, array $keys, ...$argument)
  * @method mixed             evalsha(string $script, int $numkeys, string ...$keyOrArg = null)
+ * @method mixed             evalsha_ro(string $sha1, array $keys, ...$argument)
  * @method mixed             script($subcommand, $argument = null)
  * @method mixed             auth(string $password)
  * @method string            echo(string $message)
@@ -210,7 +221,8 @@ use Predis\Response\Status;
  * @method array             geosearch(string $key, FromInterface $from, ByInterface $by, ?string $sorting = null, int $count = -1, bool $any = false, bool $withCoord = false, bool $withDist = false, bool $withHash = false)
  * @method int               geosearchstore(string $destination, string $source, FromInterface $from, ByInterface $by, ?string $sorting = null, int $count = -1, bool $any = false, bool $storeDist = false)
  *
- * @author Daniele Alessandri <suppakilla@gmail.com>
+ * Container commands
+ * @property FunctionContainer $function
  */
 interface ClientInterface
 {
@@ -253,7 +265,7 @@ interface ClientInterface
      *
      * @return CommandInterface
      */
-    public function createCommand($method, $arguments = array());
+    public function createCommand($method, $arguments = []);
 
     /**
      * Executes the specified Redis command.